refactor: 저장 버튼 이벤트 핸들러 글로벌로 등록
