+++ src/js/ui/template.js
@@ -19,31 +19,33 @@ const getFoundResultTemplate = items => {
       const { publishedAt, channelId, title, thumbnails, channelTitle } =
         item.snippet;
       return `
-      <li class="video-item" data-video-id=${item.id.videoId}>
-        <a
-          href="https://www.youtube.com/watch?v=${item.id.videoId}"
-          target="_blank"
-        >
-          <img
-            src=${thumbnails.high.url}
-            alt="video-item-thumbnail" class="video-item__thumbnail">
-        </a>
-        <a href="https://www.youtube.com/watch?v=${item.id.videoId}"
-          target="_blank"
-        >
-          <h4 class="video-item__title">${title}</h4>
-        </a>
-        <a href="https://www.youtube.com/channel/${channelId}" target="_blank">
-          <p class="video-item__channel-name">${channelTitle}</p>
-        </a>
-        <p class="video-item__published-date">
-          ${convertToKoreaLocaleDate(publishedAt)}
-        </p>
-        <button class="video-item__save-button button" ${
-          saveDatas.includes(item.id.videoId) ? 'hidden' : ''
-        }>⬇ 저장</button>
-      </li>
-    `;
+        <li class="video-item" data-video-id=${item.id.videoId}>
+          <a
+            href="https://www.youtube.com/watch?v=${item.id.videoId}"
+            target="_blank"
+          >
+            <img
+              src=${thumbnails.high.url}
+              alt="video-item-thumbnail" class="video-item__thumbnail">
+          </a>
+          <a href="https://www.youtube.com/watch?v=${item.id.videoId}"
+            target="_blank"
+          >
+            <h4 class="video-item__title">${title}</h4>
+          </a>
+          <a href="https://www.youtube.com/channel/${channelId}" target="_blank">
+            <p class="video-item__channel-name">${channelTitle}</p>
+          </a>
+          <p class="video-item__published-date">
+            ${convertToKoreaLocaleDate(publishedAt)}
+          </p>
+          <button
+            class="video-item__save-button button"
+            onclick="saveVideo(event);"
+            ${saveDatas.includes(item.id.videoId) ? 'hidden' : ''}
+          >⬇ 저장</button>
+        </li>
+      `;
     })
     .join('');
   return resultTemplate;
